Sum rounded subtotals instead of rounding only the final total
Even if the rounding for subtotal is set to two while price unit is 5,
OpenERP will keep 5 digits everywhere until it gets to the database or
the view. For this reason, we have to do roundings by ourselves in some
places.

In this case, OpenERP seems to be summing the rounded subtotals instead
of doing a sum on all lines and then rounding.