[10.0][FIX] avoid reports with empty rows
This is a fix to avoid having empty rows on the QWeb sale order
report that relate to the formatted_note field. If the field has no
content, it shouldn't be displayed. Also note that the default
formatted_note content is <p><br></p> and so we also need to prevent
printing when the field has that value.