[FIX] sale_report_delivered: Fix unsigned_price_subtotal and unsigned_product_uom_qty
When revaluation of Valuation Layers happens, another valuation is created linked to the Stock Move and the report lines were duplicated.
Using the quantity on the Valuation Layer (if exists) before taking product_uom_qty, adjust the real Quantity delivered and Price delivered on each Valuation Layer line of the report.