[FIX] sale_report_delivered: Exclude done moves with 0.0 quantities done
Done moves are splitted in 2 when quantity done exceedes quantity reserved.
When valuation layer is confirmed, only is linked to the first move, using the quantity done and not the product_uom_qty
So, if these qty done 0.0 moves are excluded, the one with all quantity done will be propagated across the report.
Also, use ABSolute values for unsigned values in sub_select because in select query will be multiplied by signed_qty.