[FIX] sale_layout_category_hide_detail: compatiblity with `sale_order_line_no_print`
The new `sale_order_line_no_print` also searches for the same nodes. Since this module was altering the `if` definition in the same views, you couldn't install both modules at the same time.

This way of overriding the views produces the same result, but doesn't alter the `if` definition, so it is compatible with the other modules.

@moduon MT-7780